feat: Add TableB component to widgets
This commit adds the TableB component to the widgets module. The TableB component was missing from the exports in the index file, causing it to be unavailable for use in other parts of the application. This commit fixes that issue by including the export statement for the TableB component in the index file.